+++ tests/training/test_training_and_offline_diags.py
@@ -9,9 +9,9 @@ import yaml
 
 import diagnostics_utils as utils
 import synth
-from fv3net.regression import shared
-from fv3net.regression.sklearn import train
-from fv3net.regression.sklearn._mapper import SklearnPredictionMapper
+from fv3fit import _shared as shared
+from fv3fit.sklearn._train import train_model
+from offline_ml_diags._mapper import SklearnPredictionMapper
 from loaders import SAMPLE_DIM_NAME, batches, mappers
 from offline_ml_diags.compute_diags import (
     _average_metrics_dict,
@@ -134,7 +134,7 @@ def _one_step_train_config(datadir_module):
         os.path.join(str(datadir_module), "train_sklearn_model_onestep_source.yml"), "r"
     ) as f:
         config = yaml.safe_load(f)
-    return train.ModelTrainingConfig(**config)
+    return shared.ModelTrainingConfig(**config)
 
 
 @pytest.fixture
@@ -147,7 +147,7 @@ def _nudging_train_config(datadir_module):
         os.path.join(str(datadir_module), "train_sklearn_model_nudged_source.yml"), "r"
     ) as f:
         config = yaml.safe_load(f)
-    return train.ModelTrainingConfig(**config)
+    return shared.ModelTrainingConfig(**config)
 
 
 @pytest.fixture
@@ -160,7 +160,7 @@ def _fine_res_train_config(datadir_module):
         os.path.join(str(datadir_module), "train_sklearn_model_fineres_source.yml"), "r"
     ) as f:
         config = yaml.safe_load(f)
-    return train.ModelTrainingConfig(**config)
+    return shared.ModelTrainingConfig(**config)
 
 
 @pytest.fixture
@@ -190,7 +190,7 @@ def training_batches(data_source_name, data_source_path, data_source_train_confi
 def test_sklearn_regression(training_batches, data_source_train_config):
 
     assert len(training_batches) == 2
-    wrapper = train.train_model(training_batches, data_source_train_config)
+    wrapper = train_model(training_batches, data_source_train_config)
     assert wrapper.model.n_estimators == 2